requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
packaging==20.8
pkg-resources==0.0.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-html==3.1.1
pytest-metadata==1.11.0
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0
